Removed abstract class for requests
