+++ src/main/java/com/progress/tracking/rest/request/SearchCourseRequest.java
@@ -1,9 +1,27 @@
 package com.progress.tracking.rest.request;
 
-public class SearchCourseRequest extends AbstractRequest {
+public class SearchCourseRequest {
+    private String udemyClientId;
+    private String udemyClientSecret;
     private String course;
     private Integer max;
     private Integer page;
+
+    public String getUdemyClientId() {
+        return udemyClientId;
+    }
+
+    public void setUdemyClientId(String udemyClientId) {
+        this.udemyClientId = udemyClientId;
+    }
+
+    public String getUdemyClientSecret() {
+        return udemyClientSecret;
+    }
+
+    public void setUdemyClientSecret(String udemyClientSecret) {
+        this.udemyClientSecret = udemyClientSecret;
+    }
 
     public String getCourse() {
         return course;
